feat: externalize the page layout left panel
