+++ packages/twenty-front/src/modules/page-layout/components/PageLayoutLeftPanel.tsx
@@ -1,0 +1,35 @@
+import { SummaryCard } from '@/object-record/record-show/components/SummaryCard';
+import { PageLayoutGridLayout } from '@/page-layout/components/PageLayoutGridLayout';
+import { useCurrentPageLayout } from '@/page-layout/hooks/useCurrentPageLayout';
+import { useLayoutRenderingContext } from '@/ui/layout/contexts/LayoutRenderingContext';
+import { useTargetRecord } from '@/ui/layout/contexts/useTargetRecord';
+import { ShowPageLeftContainer } from '@/ui/layout/show-page/components/ShowPageLeftContainer';
+import { PageLayoutType } from '~/generated/graphql';
+
+type PageLayoutLeftPanelProps = {
+  pinnedLeftTabId: string;
+};
+
+export const PageLayoutLeftPanel = ({
+  pinnedLeftTabId,
+}: PageLayoutLeftPanelProps) => {
+  const { currentPageLayout } = useCurrentPageLayout();
+  const targetRecordIdentifier = useTargetRecord();
+  const { isInRightDrawer } = useLayoutRenderingContext();
+
+  if (currentPageLayout?.type !== PageLayoutType.RECORD_PAGE) {
+    return null;
+  }
+
+  return (
+    <ShowPageLeftContainer>
+      <SummaryCard
+        objectNameSingular={targetRecordIdentifier.targetObjectNameSingular}
+        objectRecordId={targetRecordIdentifier.id}
+        isInRightDrawer={isInRightDrawer}
+      />
+
+      <PageLayoutGridLayout tabId={pinnedLeftTabId} />
+    </ShowPageLeftContainer>
+  );
+};
